KAAL_AI_backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pinecone
pymongo>=4.6.1
motor>=3.0.0
certifi



